requirements.txt
flask
sqlalchemy
psycopg2
pyjwt
bcrypt
